refactor: Grouped files for create-file-path-maps (ember-addon)
