+++ src/migration/ember-addon/steps/create-file-path-maps.ts
@@ -1,9 +1,9 @@
 import type { FilePathMap } from '@codemod-utils/files';
 
 import type { Options } from '../../../types/index.js';
-import { migrationStrategyForAddonFolder } from '../addon/index.js';
-import { migrationStrategyForAppFolder } from '../app/index.js';
-import { migrationStrategyForTestsFolder } from '../tests/index.js';
+import { migrationStrategyForAddonFolder } from './create-file-path-maps/addon/index.js';
+import { migrationStrategyForAppFolder } from './create-file-path-maps/app/index.js';
+import { migrationStrategyForTestsFolder } from './create-file-path-maps/tests/index.js';
 
 type MigrationStrategies = {
   addon: FilePathMap;
