+++ src/migration/ember-addon/steps/create-file-path-maps/addon/component-templates.ts
@@ -1,0 +1,30 @@
+import { findFiles, renamePathByFile } from '@codemod-utils/files';
+
+import type {
+  FilePathMapEntries,
+  Options,
+} from '../../../../../types/index.js';
+
+export function migrationStrategyForComponentTemplates(
+  options: Options,
+): FilePathMapEntries {
+  const { projectRoot } = options;
+
+  const filePaths = findFiles('addon/components/**/template.hbs', {
+    projectRoot,
+  });
+
+  return filePaths.map((oldFilePath) => {
+    const newFilePath = renamePathByFile(oldFilePath, {
+      find: {
+        directory: 'addon/components',
+        file: 'template',
+      },
+      replace: (key: string) => {
+        return `addon/components/${key}`;
+      },
+    });
+
+    return [oldFilePath, newFilePath];
+  });
+}
