+++ src/migration/ember-addon/steps/create-file-path-maps/app/component-templates.ts
@@ -1,0 +1,30 @@
+import { findFiles, renamePathByFile } from '@codemod-utils/files';
+
+import type {
+  FilePathMapEntries,
+  Options,
+} from '../../../../../types/index.js';
+
+export function migrationStrategyForComponentTemplates(
+  options: Options,
+): FilePathMapEntries {
+  const { projectRoot } = options;
+
+  const filePaths = findFiles('app/components/**/template.js', {
+    projectRoot,
+  });
+
+  return filePaths.map((oldFilePath) => {
+    const newFilePath = renamePathByFile(oldFilePath, {
+      find: {
+        directory: 'app/components',
+        file: 'template',
+      },
+      replace: (key: string) => {
+        return `app/components/${key}`;
+      },
+    });
+
+    return [oldFilePath, newFilePath];
+  });
+}
