+++ src/migration/ember-addon/steps/create-file-path-maps/tests/index.ts
@@ -1,0 +1,6 @@
+import type { FilePathMap, Options } from '../../../../../types/index.js';
+import { migrationStrategyForComponents } from './components.js';
+
+export function migrationStrategyForTestsFolder(options: Options): FilePathMap {
+  return new Map([...migrationStrategyForComponents(options)]);
+}
